+++ tests/minimal_project/build.py
@@ -14,6 +14,7 @@ libmath = project.library("math", sources=["factorial.cpp", "fibonacci.cpp"], li
 libtest = project.library("test", sources=["factorial.cpp", "fibonacci.cpp"], libs=["stdc++"])
 install_dir = os.path.join(CURDIR, "build")
 project.install(libtest, dir=install_dir)
+project.install("utils.hpp", dir=install_dir)
 test = project.executable("test", sources=["test.cpp"], libs=["stdc++", libtest])
 
 srbuild.cli(project)